Update home.js
entered code for the function getMostCommonGenres and created a helper function named _topFiveMost Common. Tested code in Qualified. Kept getting an error for "unexpected ending" and realized that I forgot to close out the helper function with a " } " closing end of curly brackets. Updated code in GitHub and in Qualified, reran tests, and everything passed.